adding authentication to edit & delete post

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,15 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -69,6 +78,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Post');
+        }
+
         return view('posts.edit')->with('post',$post);
     }
 
@@ -102,6 +116,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Post');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success','Postingan Dihapus !');
     }
